# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: S-PAY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/balance:
    get:
      operationId: getWalletBalance
      tags: [wallet]
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallet/transactions:
    get:
      operationId: getWalletTransactions
      tags: [wallet]
      summary: Get wallet transaction history
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /wallet/send:
    post:
      operationId: sendMoney
      tags: [wallet]
      summary: Send money to another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMoneyRequest"
      responses:
        "200":
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient balance or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/add-funds:
    post:
      operationId: addFunds
      tags: [wallet]
      summary: Add funds to wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFundsRequest"
      responses:
        "200":
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddFundsResponse"
...
  /admin/fees:
    get:
      operationId: getFeeSchedule
      summary: Read the platform fee schedule (provider cost + S-PAY margin = user price)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current fee schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeSchedule"
    put:
      operationId: updateFeeSchedule
      summary: Update the platform fee schedule (applies to new quotes immediately, no deploy)
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeSchedule"
      responses:
        "200":
          description: Updated fee schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeSchedule"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    WalletBalance:
      type: object
      properties:
        balance:
          type: number
        availableBalance:
          type: number
        currency:
          type: string
        celoAddress:
          type: string
        usdcBalance:
          type: number
        usdtBalance:
          type: number
      required:
        - balance
        - availableBalance
        - currency
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [send, receive, withdraw, recharge, payment]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        counterparty:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - currency
        - description
        - status
        - createdAt
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - transactions
        - total
        - limit
        - offset
    SendMoneyRequest:
      type: object
      properties:
        recipientPhone:
          type: string
        recipientAddress:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: USDC
        note:
          type: string
      required:
        - amount
        - currency
    AddFundsRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
          default: USD
        method:
          type: string
          enum: [bank_transfer, card]
      required:
        - amount
        - currency
        - method
    AddFundsResponse:
      type: object
      properties:
        depositId:
          type: string
        instructions:
          type: string
        accountNumber:
          type: string
        routingNumber:
          type: string
        amount:
          type: number
        currency:
          type: string
      required:
        - depositId
        - instructions
        - amount
        - currency
...
    FeeSchedule:
      type: object
      properties:
        withdrawalFeePercent:
          type: number
          description: User-facing withdrawal fee in percent (covers Noah's cost + S-PAY margin)
        withdrawalFeeMin:
          type: number
          description: Minimum withdrawal fee in USD
        cardIssuanceFee:
          type: number
          description: One-time virtual card creation fee in USD (covers Stripe's cost + margin)
        p2pFeePercent:
          type: number
          description: Internal S-PAY-to-S-PAY transfer fee in percent (0 = free, growth-friendly)
      required:
        - withdrawalFeePercent
        - withdrawalFeeMin
        - cardIssuanceFee
        - p2pFeePercent
...